refactor: move sandbox types
