+++ src/config/sandboxProcessCache.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { Duration } from '@salesforce/kit';
-import { SandboxProcessObject, SandboxRequest } from '../org';
+import { SandboxProcessObject, SandboxRequest } from '../org/sandbox/types';
 import { Global } from '../global';
 import { TTLConfig } from './ttlConfig';
